requirements.txt
fastapi==0.115.0
uvicorn==0.31.1
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
aiofiles==24.1.0
pyyaml==6.0
sentence-transformers
scikit-learn
pydantic>=2.0
